requirements.txt
Office365-REST-Python-Client==2.3.1
temp==2020.7.2
pathlib2==2.3.5
pandas>=1.2.0